TW-1584: message_download_content logic moved to mixin
(cherry picked from commit e8f98c6679e5e3f5d586f0b1f26a58408b40733e)
