+++ lib/pages/chat/events/message_download_content.dart
@@ -1,20 +1,10 @@
-import 'dart:async';
-import 'dart:io';
-
-import 'package:dartz/dartz.dart' hide State, OpenFile;
-import 'package:fluffychat/app_state/failure.dart';
-import 'package:fluffychat/app_state/success.dart';
-import 'package:fluffychat/di/global/get_it_initializer.dart';
 import 'package:fluffychat/presentation/model/chat/downloading_state_presentation_model.dart';
-import 'package:fluffychat/utils/exception/downloading_exception.dart';
-import 'package:fluffychat/utils/manager/download_manager/download_file_state.dart';
-import 'package:fluffychat/utils/manager/download_manager/download_manager.dart';
-import 'package:fluffychat/utils/matrix_sdk_extensions/download_file_extension.dart';
 import 'package:fluffychat/utils/matrix_sdk_extensions/event_extension.dart';
-import 'package:fluffychat/utils/manager/storage_directory_manager.dart';
 import 'package:fluffychat/widgets/file_widget/download_file_tile_widget.dart';
+import 'package:fluffychat/widgets/file_widget/downloading_file_tile_widget.dart';
 import 'package:fluffychat/widgets/file_widget/file_tile_widget.dart';
 import 'package:fluffychat/widgets/file_widget/message_file_tile_style.dart';
+import 'package:fluffychat/widgets/mixins/download_file_on_mobile_mixin.dart';
 import 'package:fluffychat/widgets/mixins/handle_download_and_preview_file_mixin.dart';
 import 'package:flutter/material.dart';
 import 'package:matrix/matrix.dart';
@@ -35,112 +25,11 @@ class MessageDownloadContent extends StatefulWidget {
 }
 
 class _MessageDownloadContentState extends State<MessageDownloadContent>
-    with HandleDownloadAndPreviewFileMixin {
-  final downloadManager = getIt.get<DownloadManager>();
-
-  final downloadFileStateNotifier = ValueNotifier<DownloadPresentationState>(
-    const NotDownloadPresentationState(),
-  );
-
-  StreamSubscription<Either<Failure, Success>>? streamSubscription;
-
+    with
+        HandleDownloadAndPreviewFileMixin,
+        DownloadFileOnMobileMixin<MessageDownloadContent> {
   @override
-  void initState() {
-    super.initState();
-    checkDownloadFileState();
-  }
-
-  void checkDownloadFileState() async {
-    checkFileExistInMemory();
-    await checkFileInDownloadsInApp();
-
-    _trySetupDownloadingStreamSubcription();
-    if (streamSubscription != null) {
-      downloadFileStateNotifier.value = const DownloadingPresentationState();
-    }
-  }
-
-  void checkFileExistInMemory() {
-    final filePathInMem = widget.event.getFilePathFromMem();
-    if (filePathInMem?.isNotEmpty == true) {
-      downloadFileStateNotifier.value = DownloadedPresentationState(
-        filePath: filePathInMem!,
-      );
-      return;
-    }
-  }
-
-  Future<void> checkFileInDownloadsInApp() async {
-    final filePath =
-        await StorageDirectoryManager.instance.getFilePathInAppDownloads(
-      eventId: widget.event.eventId,
-      fileName: widget.event.filename,
-    );
-    final file = File(filePath);
-    if (await file.exists() &&
-        await file.length() == widget.event.getFileSize()) {
-      downloadFileStateNotifier.value = DownloadedPresentationState(
-        filePath: filePath,
-      );
-      return;
-    }
-  }
-
-  void _trySetupDownloadingStreamSubcription() {
-    streamSubscription = downloadManager
-        .getDownloadStateStream(widget.event.eventId)
-        ?.listen(setupDownloadingProcess);
-  }
-
-  void setupDownloadingProcess(Either<Failure, Success> event) {
-    event.fold(
-      (failure) {
-        Logs().e('MessageDownloadContent::onDownloadingProcess(): $failure');
-        if (failure is DownloadFileFailureState &&
-            failure.exception is CancelDownloadingException) {
-          downloadFileStateNotifier.value =
-              const NotDownloadPresentationState();
-        } else {
-          downloadFileStateNotifier.value =
-              DownloadErrorPresentationState(error: failure);
-        }
-        streamSubscription?.cancel();
-      },
-      (success) {
-        if (success is DownloadingFileState) {
-          if (success.total != 0) {
-            downloadFileStateNotifier.value = DownloadingPresentationState(
-              receive: success.receive,
-              total: success.total,
-            );
-          }
-        } else if (success is DownloadNativeFileSuccessState) {
-          downloadFileStateNotifier.value = DownloadedPresentationState(
-            filePath: success.filePath,
-          );
-        }
-      },
-    );
-  }
-
-  void onDownloadFileTap() async {
-    await checkFileInDownloadsInApp();
-    if (downloadFileStateNotifier.value is DownloadedPresentationState) {
-      return;
-    }
-    downloadFileStateNotifier.value = const DownloadingPresentationState();
-    downloadManager.download(
-      event: widget.event,
-    );
-    _trySetupDownloadingStreamSubcription();
-  }
-
-  @override
-  void dispose() {
-    streamSubscription?.cancel();
-    downloadFileStateNotifier.dispose();
-    super.dispose();
-  }
+  Event get event => widget.event;
 
   @override
   Widget build(BuildContext context) {
@@ -187,8 +76,8 @@ class _MessageDownloadContentState extends State<MessageDownloadContent>
         }
 
         return InkWell(
-          onTap: onDownloadFileTap,
-          child: DownloadFileTileWidget(
+          onTap: () => onDownloadFileTap(),
+          child: DownloadingFileTileWidget(
             mimeType: widget.event.mimeType,
             fileType: filetype,
             filename: filename,
