+++ lib/widgets/file_widget/file_tile_widget.dart
@@ -1,151 +1,58 @@
-import 'dart:typed_data';
-
 import 'package:fluffychat/presentation/model/chat/downloading_state_presentation_model.dart';
 import 'package:fluffychat/utils/extension/mime_type_extension.dart';
 import 'package:fluffychat/utils/matrix_sdk_extensions/int_extension.dart';
-import 'package:fluffychat/utils/string_extension.dart';
+import 'package:fluffychat/widgets/file_widget/base_file_tile_widget.dart';
 import 'package:fluffychat/widgets/file_widget/file_tile_widget_style.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 
-class FileTileWidget extends StatelessWidget {
-  const FileTileWidget({
+class FileTileWidget extends BaseFileTileWidget {
+  FileTileWidget({
     super.key,
-    required this.mimeType,
-    required this.filename,
-    this.fileType,
-    this.highlightText,
-    this.sizeString,
-    this.backgroundColor,
-    this.fileTileIcon,
-    this.imageBytes,
-    this.style = const FileTileWidgetStyle(),
-  });
-
-  final TwakeMimeType mimeType;
-  final String filename;
-  final String? highlightText;
-  final String? sizeString;
-  final Color? backgroundColor;
-  final String? fileType;
-  final Uint8List? imageBytes;
-  final String? fileTileIcon;
-  final FileTileWidgetStyle style;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: style.paddingFileTileAll,
-      decoration: ShapeDecoration(
-        color: backgroundColor ?? style.backgroundColor,
-        shape: RoundedRectangleBorder(
-          borderRadius: style.borderRadius,
-        ),
-      ),
-      child: Row(
-        crossAxisAlignment: style.crossAxisAlignment,
-        children: [
-          if (imageBytes != null)
-            Padding(
-              padding: style.imagePadding,
-              child: ClipRRect(
-                borderRadius: style.borderRadius,
-                child: Image.memory(
-                  imageBytes!,
-                  width: style.imageSize,
-                  height: style.imageSize,
-                  fit: BoxFit.cover,
+    required super.mimeType,
+    required super.filename,
+    super.fileType,
+    super.highlightText,
+    super.sizeString,
+    super.backgroundColor,
+    super.fileTileIcon,
+    super.imageBytes,
+    super.style = const FileTileWidgetStyle(),
+  }) : super(
+          subTitle: (context) => Row(
+            children: [
+              if (sizeString != null)
+                TextInformationOfFile(
+                  value: sizeString,
+                  style: style.textInformationStyle(context),
+                ),
+              TextInformationOfFile(
+                value: " · ",
+                style: style.textInformationStyle(context),
+              ),
+              Flexible(
+                child: TextInformationOfFile(
+                  value: mimeType.getFileType(
+                    context,
+                    fileType: fileType,
+                  ),
+                  style: style.textInformationStyle(context),
                 ),
               ),
-            ),
-          if (imageBytes == null)
-            SvgPicture.asset(
-              fileTileIcon ?? mimeType.getIcon(fileType: fileType),
-              width: style.iconSize,
-              height: style.iconSize,
-            ),
-          style.paddingRightIcon,
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisSize: MainAxisSize.max,
-              children: [
-                const SizedBox(height: 4.0),
-                FileNameText(
-                  filename: filename,
-                  highlightText: highlightText,
-                  style: style,
-                ),
-                Row(
-                  children: [
-                    if (sizeString != null)
-                      TextInformationOfFile(
-                        value: sizeString!,
-                        style: style,
-                      ),
-                    TextInformationOfFile(
-                      value: " · ",
-                      style: style,
-                    ),
-                    Flexible(
-                      child: TextInformationOfFile(
-                        value: mimeType.getFileType(
-                          context,
-                          fileType: fileType,
-                        ),
-                        style: style,
-                      ),
-                    ),
-                  ],
-                ),
-                style.paddingBottomText,
-              ],
-            ),
+            ],
           ),
-        ],
-      ),
-    );
-  }
-}
-
-class FileNameText extends StatelessWidget {
-  const FileNameText({
-    super.key,
-    required this.filename,
-    this.highlightText,
-    this.style = const FileTileWidgetStyle(),
-  });
-
-  final String filename;
-  final String? highlightText;
-  final FileTileWidgetStyle style;
-
-  @override
-  Widget build(BuildContext context) {
-    return Text.rich(
-      TextSpan(
-        children: filename.buildHighlightTextSpans(
-          highlightText ?? '',
-          style: style.textStyle(context),
-          highlightStyle: style.highlightTextStyle(context),
-        ),
-      ),
-      maxLines: 2,
-      overflow: TextOverflow.ellipsis,
-    );
-  }
+        );
 }
 
 class TextInformationOfFile extends StatelessWidget {
   final String value;
-  final FileTileWidgetStyle style;
+  final TextStyle? style;
   final ValueNotifier<DownloadPresentationState>? downloadFileStateNotifier;
 
   const TextInformationOfFile({
     super.key,
     required this.value,
     this.downloadFileStateNotifier,
-    this.style = const FileTileWidgetStyle(),
+    this.style,
   });
 
   @override
@@ -162,7 +69,7 @@ class TextInformationOfFile extends StatelessWidget {
                   downloadFileState.total! >= IntExtension.oneKB) {
                 return Text(
                   '${downloadFileState.receive!.bytesToMB(placeDecimal: 1)} MB / ',
-                  style: style.textInformationStyle(context),
+                  style: style,
                   maxLines: 1,
                   overflow: TextOverflow.ellipsis,
                 );
@@ -172,7 +79,7 @@ class TextInformationOfFile extends StatelessWidget {
           ),
         Text(
           value,
-          style: style.textInformationStyle(context),
+          style: style,
           maxLines: 1,
           overflow: TextOverflow.ellipsis,
         ),
